Update README.md
This project is a web-based application designed for classifying food images using a pre-trained ResNet50 model. It utilizes convolutional neural networks (CNNs) for image processing and prediction. The application is implemented with Streamlit, providing an interactive interface for users to upload food images and get classification results. This project is intended for a final year academic requirement